requirements.txt
# Main dependencies
PySide6>=6.4.0
pandas>=2.0.0
numpy>=1.24.0
liac-arff>=2.5.0
matplotlib>=3.7.0